refactor: alterando tipo dos parâmetros de Integer para Double

+++ src/main/java/com/greenbeansapps/myschooltransportation/domain/usecases/UpdateStudentUseCase.java
@@ -5,5 +5,5 @@ import com.greenbeansapps.myschooltransportation.domain.entities.Student;
 import java.util.UUID;
 
 public interface UpdateStudentUseCase {
-  Student execute(UUID StudentId, String name, String school, String grade, String transportationType, Integer monthlyPayment, Integer monthlyPaymentExpiration, String shift);
+  Student execute(UUID StudentId, String name, String school, String grade, String transportationType, Double monthlyPayment, Integer monthlyPaymentExpiration, String shift);
 }
